+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.9.3] - 2019-10-01
+### Fixed
+* fix that keepRatio want to behave like sketches and illustrators. #47
+* fix keepRatio default false
+
+
 ## [0.9.2] - 2019-10-01
 ### Fixed
 * fix that bounds don't apply when snap